+++ dev/zotero/zotero-server-commands.ts
@@ -1,0 +1,4 @@
+export default class ZoteroServerCommand {
+    static readonly GET_ACTIVE_DOCUMENT = "Application_getActiveDocument";
+    static readonly GET_DOCUMENT_DATA = "Document_getDocumentData";
+}