+++ src/constants/command-keys.ts
@@ -1,0 +1,8 @@
+enum CommandKeys {
+  SHOW_PREVIEW = "cslPreview.showCslPreview",
+  REFRESH_PREVIEW = "cslPreview.refreshCslPreview",
+  SHOW_SOURCE = "cslPreview.showPreviewSource",
+  CHOOSE_LOCALE = "cslPreview.chooseLocale"
+}
+
+export default CommandKeys;