+++ src/events/listener.ts
@@ -1,0 +1,4 @@
+export default interface Listener<E> {
+  callback(event: E): void;
+  dispose(): void;
+}