+++ src/models/citable.ts
@@ -1,0 +1,5 @@
+export default interface Citable {
+  id: string;
+  type: string;
+  title: string;
+}