+++ src/types/csl-validator.d.ts
@@ -1,0 +1,3 @@
+declare function validate(input: string): string;
+
+export { validate };